(Menu) Use float for icon coordinates, makes animations look smoother

+++ menu/drivers_display/menu_display_null.c
@@ -40,7 +40,7 @@ static void menu_display_null_blend_end(void)
 }
 
 static void menu_display_null_draw(
-      unsigned x, unsigned y,
+      float x, float y,
       unsigned width, unsigned height,
       struct gfx_coords *coords,
       void *matrix_data,
